Allow setting source tag
Private trackers that use this often fill in a three letter code specific for
their tracker. Users that upload torrents can add it themselves (using
dottorrent-gui for example) or the tracker will add it and the uploader needs
to download the torrent with added tag from the tracker to start the seeding.
Often people use different a locations for uploads than the default download
location. So they will have to move files or edit download location and do a
force recheck. If it would be implemented in the normal torrent creation
process it would be much easier especially for novice users.

If if was not clear this is done to be able to cross-seed the same content
between different private trackers. If no difference is made between the
torrents the tracker addresses of the 2nd torrent will be added to the first
and the seeders of both trackers will be mixed by the client.

Fixes #1